+++ app/src/main/java/pl/lambada/songsync/ui/screens/home/components/batchDownload/RateLimitedDialog.kt
@@ -1,0 +1,31 @@
+package pl.lambada.songsync.ui.screens.home.components.batchDownload
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import pl.lambada.songsync.R
+
+@Composable
+fun RateLimitedDialog(onDismiss: () -> Unit) {
+    AlertDialog(
+        title = {
+            Text(text = stringResource(id = R.string.batch_download_lyrics))
+        },
+        text = {
+            Column {
+                Text(text = stringResource(R.string.spotify_api_rate_limit_reached))
+                Text(text = stringResource(R.string.please_try_again_later))
+                Text(text = stringResource(R.string.change_api_strategy))
+            }
+        },
+        onDismissRequest = onDismiss,
+        confirmButton = {
+            Button(onClick = onDismiss) {
+                Text(text = stringResource(id = R.string.ok))
+            }
+        }
+    )
+}
